refactor(gabinet): extract deliveries route components (closes #4454)
